requirements.txt
streamlit==1.32.0
pandas==2.2.0
folium==0.15.1
plotly==5.18.0
requests==2.31.0
Pillow==10.2.0
streamlit-folium==0.15.0
numpy==1.26.3 